Add support for resume and pause.
I've added support for a "resume" and "pause" method to be used when a server is
killed or unkilled. They currently return an error that is unused because
I haven't thought of a good use for it.

I also changed the init method to take a list of killfile path instead of
generating them internally. There is now a function that generates the default
paths of port + all.

Servicez now takes an interface and marshals it to json

I've also encapsulated the adminz stuff into an object.